# openapi.yaml
openapi: 3.0.3
info:
  title: Kloka Attendance Platform API
  version: 1.0.0
  description: Full API surface (88 paths, 109 operations) for the Kloka zero-trust workforce attendance platform.
    Generated from the locked API_DESIGN.md. Schemas are typed where the domain is built (clocking, auth)
    and generic objects where the domain is still designed; tighten per-domain as code lands.
  contact:
    name: Kloka Engineering
    email: engineering@kloka.co.za
servers:
- url: https://api.kloka.co.za/api/v1
  description: Production
- url: http://localhost:8080/api/v1
  description: Local
...
paths:
...
  /tenants:
    post:
      operationId: PostTenants
      summary: Create a tenant and issue its one-time tenant key
      description: Platform-internal. Provisions a tenant, generates its server-side JWT signing secret
        (never returned), and returns a one-time tenant API key used to scope worker/admin login via the
        X-Tenant-Key header.
      tags:
      - Tenants
      security:
      - platformKey: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateTenantRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateTenantResponse'
          headers:
            Location:
              schema:
                type: string
                format: uri
              description: URI of the created resource.
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '422':
          $ref: '#/components/responses/Unprocessable'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
        '415':
          $ref: '#/components/responses/UnsupportedMediaType'
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
    get:
      operationId: GetTenants
      summary: List accessible tenants
      tags:
      - Tenants
      security:
      - bearerAuth: []
      responses:
        '200':
          description: Paginated list
          content:
            application/json:
              schema:
                type: object
                required:
                - data
                - next_cursor
                - has_more
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TenantResponse'
                  next_cursor:
                    type: string
                    nullable: true
                  limit:
                    type: integer
                    example: 50
                  has_more:
                    type: boolean
                    description: True if more pages exist (next_cursor != null).
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
      parameters:
      - $ref: '#/components/parameters/TenantHeader'
      - $ref: '#/components/parameters/Cursor'
      - $ref: '#/components/parameters/Limit'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    sessionToken:
      type: apiKey
      in: header
      name: X-Session-Token
    refreshCookie:
      type: apiKey
      in: cookie
      name: refresh_token
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    tenantKey:
      type: apiKey
      in: header
      name: X-Tenant-Key
      description: Per-tenant API key, returned ONCE when the tenant is created. Scopes the login call
        to a tenant. This is NOT the JWT signing secret — that never leaves the server. This key only
        identifies + authorises the tenant context.
    platformKey:
      type: apiKey
      in: header
      name: X-Platform-Key
      description: Kloka-internal platform credential. Required to create tenants.
  parameters:
    TenantHeader:
      name: X-Tenant-ID
      in: header
      required: false
      description: DEPRECATED transitional header. Tenant identity will come from the JWT tenant_id claim
        once auth middleware lands; this header is the interim placeholder and must not be relied upon
        in production.
      schema:
        type: string
        format: uuid
      deprecated: true
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: UUIDv4. Makes mutating requests safe to retry / offline-replay. Accepted on all mutating
        operations for a uniform client contract; on naturally-idempotent operations (DELETE, login) the
        key is accepted but inert.
      schema:
        type: string
    Cursor:
      name: cursor
      in: query
      required: false
      description: Opaque base64 cursor from a previous response.
      schema:
        type: string
    Limit:
      name: limit
      in: query
      required: false
      schema:
        type: integer
        default: 50
        maximum: 200
  responses:
    BadRequest:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unauthorized:
      description: RFC 7807 problem
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    RateLimited:
      description: Too many requests
      headers:
        Retry-After:
          schema:
            type: integer
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Conflict:
      description: Conflict (unique constraint or idempotency-key reuse with different payload)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    Unprocessable:
      description: Business-rule violation
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
    UnsupportedMediaType:
      description: Content-Type not supported (expected application/json)
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
  schemas:
...
    Problem:
      type: object
      required:
      - type
      - title
      - status
      properties:
        type:
          type: string
          format: uri
          example: https://api.kloka.co.za/errors/validation
        title:
          type: string
          example: Validation failed
        status:
          type: integer
          example: 400
        detail:
          type: string
          example: phone must be a valid South African mobile number (+27[6-8]xxxxxxxx)
        instance:
          type: string
          example: /api/v1/auth/worker/login
        request_id:
          type: string
          example: req_01HXYZ...
        errors:
          type: array
          description: Per-field validation errors (present on 400 validation failures).
          items:
            type: object
            required:
            - field
            - message
            properties:
              field:
                type: string
                example: pin
              message:
                type: string
                example: must be 4-8 digits
...
    CreateTenantRequest:
      type: object
      required:
      - name
      additionalProperties: false
      properties:
        name:
          type: string
          minLength: 2
          maxLength: 120
          example: Acme Retail Group
        billing_email:
          type: string
          format: email
          example: ops@acme.co.za
        country_code:
          type: string
          pattern: ^[A-Z]{2}$
          example: ZA
          description: ISO 3166-1 alpha-2.
    CreateTenantResponse:
      type: object
      required:
      - tenant_id
      - name
      - tenant_key
      - created_at
      properties:
        tenant_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        tenant_key:
          type: string
          example: tk_live_8f3a...
          description: Shown ONCE. Store securely; sent as X-Tenant-Key on login. Cannot be retrieved
            again — only rotated.
        created_at:
          type: string
          format: date-time
    TenantResponse:
      type: object
      required:
      - tenant_id
      - name
      - created_at
      properties:
        tenant_id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
        country_code:
          type: string
          example: ZA
        status:
          type: string
          enum:
          - active
          - suspended
          - archived
        created_at:
          type: string
          format: date-time
...